Move the xrapp framework into its own product flavor directory
This allows for better encapsulation of the framework without affecting the rest of the project.
