+++ core/src/xrapp/AndroidManifest.xml
@@ -1,0 +1,6 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="org.godotengine.plugin.gast">
+
+    <application/>
+
+</manifest>
