+++ core/src/xrapp/assets/.gitignore
@@ -1,0 +1,2 @@
+/android/build
+/android/.build_version
