+++ core/src/xrapp/assets/import/.gdignore
@@ -1,0 +1,1 @@
+
