+++ core/src/xrapp/assets/import/touch_controller.dae-8cc7b41dd10a947c6aef463bb282c5eb.md5
@@ -1,0 +1,3 @@
+source_md5="c05e75ae3e5bae48575ebb8f71a62e90"
+dest_md5="438ee4a19d27aae9dd4c56a60c4d2dc0"
+
