+++ core/src/xrapp/assets/main.tscn
@@ -1,0 +1,89 @@
+[gd_scene load_steps=11 format=2]
+
+[ext_resource path="res://main.gd" type="Script" id=1]
+[ext_resource path="res://touch_controller.tscn" type="PackedScene" id=2]
+[ext_resource path="res://ray_cast_with_reticle.gd" type="Script" id=3]
+[ext_resource path="res://test_shader.shader" type="Shader" id=4]
+
+[sub_resource type="ShaderMaterial" id=1]
+shader = ExtResource( 4 )
+shader_param/albedo = Color( 0.929412, 0.858824, 0.858824, 1 )
+shader_param/specular = 0.5
+shader_param/metallic = 0.0
+shader_param/roughness = 1.0
+shader_param/point_size = 1.0
+shader_param/uv1_scale = Vector3( 1, 1, 1 )
+shader_param/uv1_offset = Vector3( 0, 0, 0 )
+shader_param/uv2_scale = Vector3( 1, 1, 1 )
+shader_param/uv2_offset = Vector3( 0, 0, 0 )
+
+[sub_resource type="CapsuleMesh" id=2]
+material = SubResource( 1 )
+radius = 0.004
+mid_height = 0.5
+
+[sub_resource type="SpatialMaterial" id=3]
+
+[sub_resource type="SphereMesh" id=4]
+material = SubResource( 3 )
+radius = 0.01
+height = 0.02
+
+[sub_resource type="SpatialMaterial" id=5]
+
+[sub_resource type="SphereMesh" id=6]
+material = SubResource( 5 )
+radius = 0.01
+height = 0.02
+
+[node name="Main" type="Spatial"]
+script = ExtResource( 1 )
+
+[node name="ARVROrigin" type="ARVROrigin" parent="."]
+
+[node name="ARVRCamera" type="ARVRCamera" parent="ARVROrigin"]
+transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1.32891, 0 )
+
+[node name="LeftTouchController" type="ARVRController" parent="ARVROrigin"]
+transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, -0.2, 1.727, -0.4 )
+
+[node name="touch_controller" parent="ARVROrigin/LeftTouchController" instance=ExtResource( 2 )]
+transform = Transform( -1, 0, -3.25841e-07, 0, 1, 0, 3.25841e-07, 0, -1, 0, 0, 0 )
+
+[node name="MeshInstance" type="MeshInstance" parent="ARVROrigin/LeftTouchController"]
+transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, -0.009, 0, -0.297 )
+mesh = SubResource( 2 )
+material/0 = null
+
+[node name="LeftRayCast" type="RayCast" parent="ARVROrigin/LeftTouchController/MeshInstance" groups=["gast_ray_caster"]]
+enabled = true
+cast_to = Vector3( 0, 0, -5 )
+script = ExtResource( 3 )
+
+[node name="RayReticle" type="MeshInstance" parent="ARVROrigin/LeftTouchController/MeshInstance/LeftRayCast"]
+mesh = SubResource( 4 )
+material/0 = null
+
+[node name="RightTouchController" type="ARVRController" parent="ARVROrigin"]
+transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, 0.2, 1.727, -0.4 )
+controller_id = 2
+
+[node name="touch_controller" parent="ARVROrigin/RightTouchController" instance=ExtResource( 2 )]
+transform = Transform( 1, 3.25841e-07, 0, -3.25841e-07, 1, 0, 0, 0, -1, 0, 0, 0 )
+
+[node name="MeshInstance" type="MeshInstance" parent="ARVROrigin/RightTouchController"]
+transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, 0.009, 0, -0.297 )
+mesh = SubResource( 2 )
+material/0 = null
+
+[node name="RightRayCast" type="RayCast" parent="ARVROrigin/RightTouchController/MeshInstance" groups=["gast_ray_caster"]]
+enabled = true
+cast_to = Vector3( 0, 0, -5 )
+script = ExtResource( 3 )
+
+[node name="RayReticle" type="MeshInstance" parent="ARVROrigin/RightTouchController/MeshInstance/RightRayCast"]
+mesh = SubResource( 6 )
+material/0 = null
+
+[node name="GastActivityContainer" type="Position3D" parent="."]
+transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1.6, -3 )
