+++ core/src/xrapp/assets/ray_cast_with_reticle.gd
@@ -1,0 +1,9 @@
+extends RayCast
+
+onready var ray_reticle = $RayReticle
+
+func _physics_process(delta):
+	ray_reticle.visible = is_colliding()
+	if (ray_reticle.visible):
+		ray_reticle.translation = to_local(get_collision_point())
+
