+++ samples/spatial_app/build.gradle
@@ -9,12 +9,6 @@ android {
     ndkVersion versions.ndkVersion
 
     defaultConfig {
-        // The default ignore pattern for the 'assets' directory includes hidden files and directories which are used by Godot projects.
-        // We also ignore the 'android' build directory which is generated by the Godot editor.
-        aaptOptions {
-            ignoreAssetsPattern "!.svn:!.git:!.ds_store:!*.scc:<dir>_*:!CVS:!thumbs.db:!picasa.ini:!*~:<dir>android"
-        }
-
         applicationId "org.godotengine.plugin.gast.samples.spatialapp"
         minSdkVersion versions.minSdk
         targetSdkVersion versions.targetSdk
@@ -22,13 +16,7 @@ android {
         versionName "1.0"
         multiDexEnabled true
 
-        missingDimensionStrategy 'features', 'core'
-    }
-
-    sourceSets {
-        main {
-            assets.srcDirs += ['godot/']
-        }
+        missingDimensionStrategy 'features', 'xrapp'
     }
 
     compileOptions {
