--- samples/spatial_app/godot/.gitignore
@@ -1,2 +1,0 @@
-/android/build
-/android/.build_version
