--- samples/spatial_app/godot/.import/touch_controller.dae-8cc7b41dd10a947c6aef463bb282c5eb.md5
@@ -1,3 +1,0 @@
-source_md5="a89ba2b1799f71e9f8cc54f21e9108d0"
-dest_md5="438ee4a19d27aae9dd4c56a60c4d2dc0"
-
