--- samples/spatial_app/godot/ray_cast_with_reticle.gd
@@ -1,9 +1,0 @@
-extends RayCast
-
-onready var ray_reticle = $RayReticle
-
-func _physics_process(delta):
-	ray_reticle.visible = is_colliding()
-	if (ray_reticle.visible):
-		ray_reticle.translation = to_local(get_collision_point())
-
