--- samples/spatial_app/godot/touch_controller.tscn
@@ -1,5 +1,0 @@
-[gd_scene load_steps=2 format=2]
-
-[ext_resource path="res://touch_controller.dae" type="PackedScene" id=1]
-
-[node name="touch_controller" instance=ExtResource( 1 )]
