+++ samples/spatial_app/src/main/java/org/godotengine/plugin/gast/samples/spatialapp/MainActivity.kt
@@ -1,7 +1,7 @@
 package org.godotengine.plugin.gast.samples.spatialapp
 
 import android.os.Bundle
-import org.godotengine.plugin.gast.app.GastActivity
+import org.godotengine.plugin.gast.xrapp.GastActivity
 
 class MainActivity : GastActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
