+++ samples/spatial_app/src/main/res/layout/activity_main.xml
@@ -2,9 +2,10 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_width="800dp"
+    android:layout_height="500dp"
     android:background="@android:color/white"
+    android:layout_gravity="center"
     tools:context=".MainActivity">
 
     <TextView
